jwt 추가 후 로그인세션 만료시 예외처리 try catch error 처리

+++ controller/LikeController.js
@@ -6,38 +6,64 @@ require("dotenv").config();
 
 const addLike = (req, res) => {
   const book_id = req.params.id;
-  const authorization = ensureAuthorization(req);
+  const authorization = ensureAuthorization(req, res);
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요.",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "잘못된 토큰입니다.",
+    });
+  } else {
+    let sql = "INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?);";
+    let values = [authorization.id, book_id];
+    connection.query(sql, values, (err, result) => {
+      if (err) {
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
 
-  let sql = "INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?);";
-  let values = [authorization.id, book_id];
-  connection.query(sql, values, (err, result) => {
-    if (err) {
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
-
-    return res.status(StatusCodes.OK).json(result);
-  });
+      return res.status(StatusCodes.OK).json(result);
+    });
+  }
 };
 
 const removeLike = (req, res) => {
   const book_id = req.params.id;
-  const authorization = ensureAuthorization(req);
-  let sql = "DELETE FROM likes WHERE user_id=? AND liked_book_id=?";
+  const authorization = ensureAuthorization(req, res);
 
-  let values = [authorization.id, book_id];
-  connection.query(sql, values, (err, result) => {
-    if (err) {
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요.",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "잘못된 토큰입니다.",
+    });
+  } else {
+    let sql = "DELETE FROM likes WHERE user_id=? AND liked_book_id=?";
 
-    return res.status(StatusCodes.OK).json(result);
-  });
+    let values = [authorization.id, book_id];
+    connection.query(sql, values, (err, result) => {
+      if (err) {
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
+
+      return res.status(StatusCodes.OK).json(result);
+    });
+  }
 };
+const ensureAuthorization = (req, res) => {
+  try {
+    let receivedJwt = req.headers["authorization"];
+    let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
 
-const ensureAuthorization = (req) => {
-  let receivedJwt = req.headers["authorization"];
-  let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
-  return decodedJwt;
+    return decodedJwt;
+  } catch (err) {
+    console.log(err.name);
+    console.log(err.message);
+    return err;
+  }
 };
 
 module.exports = { addLike, removeLike };
